pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "graphrag"
version = "0.1.0"
description = "Graph-aware repository indexer with MCP interface for LLM tooling."
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
  { name = "GraphRAG Authors" }
]
dependencies = [
  "typer>=0.12,<0.14",
  "rich>=13.0,<14.0",
  "pydantic>=2.6,<3.0",
  "PyYAML>=6.0,<7.0",
  "tree_sitter>=0.23,<0.24",
  "tree-sitter-swift>=0.0.1,<0.1.0",
  "GitPython>=3.1,<4.0",
  "mcp>=0.9.0,<1.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9.0",
  "pytest-cov>=5.0,<6.0"
]

[project.scripts]
graphrag = "graphrag.cli.main:app"
graphrag-mcp = "graphrag.mcp.server:run_server"

[tool.pytest.ini_options]
addopts = "-q"
pythonpath = [
  "src"
]

